+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
  ### Changed
 
+- Migrate to ESLint v9
 - Update dependencies
 
 ## [4.5.0] - 2024-12-10